fix(cloud): normalize hosted persona examples (#19755)
