+++ packages/cloud/shared/src/lib/utils/default-eliza-character.test.ts
@@ -88,18 +88,10 @@ describe("getDefaultElizaCharacterData", () => {
     expect(defaultAgent.character.style).toEqual(character.style);
     expect(defaultAgent.character.postExamples).toEqual(character.post_examples);
 
-    const hostedExamples = defaultAgent.character.messageExamples.map((group) =>
-      group.map((turn) => ({
-        name:
-          turn.user === "{{agentName}}"
-            ? character.name
-            : turn.user === "{{user1}}"
-              ? "{{name1}}"
-              : turn.user,
-        content: { text: turn.content.text },
-      })),
-    );
-    expect(hostedExamples).toEqual(character.message_examples);
+    expect(defaultAgent.character.messageExamples).toEqual(character.message_examples);
+    expect(defaultAgent.character.messageExamples[0]?.[0]?.name).toBe("{{name1}}");
+    expect(defaultAgent.character.messageExamples[0]?.[1]?.name).toBe(character.name);
+    expect(defaultAgent.character.messageExamples[0]?.[0]).not.toHaveProperty("user");
   });
 });
